split off calibre database class
